requirements.txt
astroid==2.4.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.5.2
python-aiml==0.9.3
six==1.14.0
toml==0.10.1
wrapt==1.12.1
